+++ apps/src/code-studio/components/SmallFooter.jsx
